Move config methods to new BuildRunnerConfig class
